cad-pulse: mute using UCM controls
On the PinePhone, muting the microphone using PulseAudio won't work as 
the audio is fully routed in hardware (and PA is not that good regarding 
HW controls support).

The best option we have is to directly call alsa in order to toggle the 
corresponding HW switch, effectively muting the microphone.